Refactored to remove files
