+++ src/components/Timeline.vue
@@ -40,7 +40,7 @@
 <script lang="ts">
 import { Vue, Component, Prop, Watch, Mixins, Inject } from 'vue-property-decorator';
 import { ResponsiveMixin, Directions } from '@/modules/resize';
-import { Button } from '@/keys';
+import { Button } from '@/utils';
 import { range, Nullable } from '@/utils';
 import Progression from '@/components/Progression.vue';
