--- src/keys.ts
@@ -1,12 +1,0 @@
-export const Keys = {
-  SHIFT: 16,
-  DELETE: 46,
-  BACKSPACE: 8,
-};
-
-
-export const Button = {
-  LEFT: 0,
-  MIDDLE: 1,
-  RIGHT: 2,
-};
