+++ src/utils.ts
@@ -107,3 +107,17 @@ export const Nullable = (o: { new(): object }) => {
     },
   };
 };
+
+
+export const Keys = {
+  SHIFT: 16,
+  DELETE: 46,
+  BACKSPACE: 8,
+};
+
+
+export const Button = {
+  LEFT: 0,
+  MIDDLE: 1,
+  RIGHT: 2,
+};
